+++ packages/react-components/src/action-form/fields/FormFieldRenderer.tsx
@@ -16,15 +16,13 @@
 
 import React, { memo } from "react";
 import { FormField } from "../FormField.js";
-import type { Option, RendererFieldDefinition } from "../FormFieldApi.js";
+import type { RendererFieldDefinition } from "../FormFieldApi.js";
+import { CustomField } from "./CustomField.js";
 import { DatetimePickerField } from "./DatetimePickerField.js";
 import { DropdownField } from "./DropdownField.js";
 import { RadioButtonsField } from "./RadioButtonsField.js";
 import { TextAreaField } from "./TextAreaField.js";
 import { TextInputField } from "./TextInputField.js";
-
-const EMPTY_ITEMS: unknown[] = [];
-const EMPTY_OPTIONS: Option<never>[] = [];
 
 export interface FormFieldRendererProps {
   fieldDefinition: RendererFieldDefinition;
@@ -81,15 +79,12 @@ function renderFieldComponent(
         />
       );
     case "DROPDOWN": {
-      const { items = EMPTY_ITEMS, ...dropdownProps } =
-        fieldDefinition.fieldComponentProps ?? {};
       return (
         <DropdownField
           value={value}
           onChange={onChange}
-          items={items}
           placeholder={fieldDefinition.placeholder}
-          {...dropdownProps}
+          {...fieldDefinition.fieldComponentProps}
         />
       );
     }
@@ -104,23 +99,27 @@ function renderFieldComponent(
           {...fieldDefinition.fieldComponentProps}
         />
       );
-    case "RADIO_BUTTONS": {
-      const { options = EMPTY_OPTIONS, ...radioProps } =
-        fieldDefinition.fieldComponentProps ?? {};
+    case "RADIO_BUTTONS":
       return (
         <RadioButtonsField
           id={fieldDefinition.fieldKey}
           value={value}
           onChange={onChange}
-          options={options}
-          {...radioProps}
+          {...fieldDefinition.fieldComponentProps}
         />
       );
-    }
+    case "CUSTOM":
+      return (
+        <CustomField
+          id={fieldDefinition.fieldKey}
+          value={value}
+          onChange={onChange}
+          {...fieldDefinition.fieldComponentProps}
+        />
+      );
     case "NUMBER_INPUT":
     case "FILE_PICKER":
     case "OBJECT_SET":
-    case "CUSTOM":
       return <div>Unsupported field type: {fieldDefinition.fieldComponent}
       </div>;
     default: