+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @osdk/widget.api
 
+## 3.5.0-beta.29
+
+### Patch Changes
+
+- Updated dependencies [4a856cb]
+  - @osdk/api@2.8.0-beta.29
+
 ## 3.5.0-beta.28
 
 ### Patch Changes